Time: 6 ms (6.93%), Space: 6.1 MB (70.95%) - LeetHub

+++ 1291-sequential-digits/1291-sequential-digits.cpp
@@ -29,7 +29,7 @@ public:
         while(stol(s)<=high)
         {
             
-            if(stoi(s)>=low)
+            if(stol(s)>=low)
                 ans.push_back(stol(s));
             
             //cout<<"here"<<endl;
